[TAN-536] Introduce useGraphDataUnits to reduce duplicated code
