+++ front/app/api/graph_data_units/useGraphDataUnits.ts
@@ -1,0 +1,49 @@
+import { useNode } from '@craftjs/core';
+import useGraphDataUnitsPublished from 'api/graph_data_units/useGraphDataUnitsPublished';
+import { useReportContext } from 'containers/Admin/reporting/context/ReportContext';
+
+import { useLocation } from 'react-router-dom';
+import { isPage } from 'utils/helperUtils';
+
+import { BaseResponseData } from 'utils/cl-react-query/fetcher';
+import useGraphDataUnitsLive from './useGraphDataUnitsLive';
+import { PropsLive, ResolvedName } from './types';
+
+interface Props {
+  resolvedName: ResolvedName;
+  queryParameters: PropsLive;
+  onSuccess?: () => void;
+}
+
+const useGraphDataUnits = <Response extends BaseResponseData>({
+  resolvedName,
+  queryParameters,
+  onSuccess,
+}: Props) => {
+  const { pathname } = useLocation();
+  const { id: graphId } = useNode();
+  const isAdminPage = isPage('admin', pathname);
+  const { reportId } = useReportContext();
+
+  const { data: analyticsLive } = useGraphDataUnitsLive<Response>(
+    {
+      resolvedName,
+      props: queryParameters,
+    },
+    { enabled: isAdminPage, onSuccess }
+  );
+
+  const { data: analyticsPublished } = useGraphDataUnitsPublished<Response>(
+    {
+      reportId,
+      graphId,
+    },
+    { enabled: !isAdminPage }
+  );
+
+  const analytics = analyticsLive ?? analyticsPublished;
+
+  return analytics;
+};
+
+export default useGraphDataUnits;
