+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/PostsByTimeWidget/usePostsByTime/index.ts
@@ -1,13 +1,7 @@
 import { useMemo, useState } from 'react';
 
 // hooks
-import { useNode } from '@craftjs/core';
-import useLivePostsByTime from 'components/admin/GraphCards/PostsByTimeCard/usePostsByTime/useLivePostsByTime';
-import useGraphDataUnitsPublished from 'api/graph_data_units/useGraphDataUnitsPublished';
-import { useReportContext } from 'containers/Admin/reporting/context/ReportContext';
-
-// routing
-import { useLocation } from 'react-router-dom';
+import useGraphDataUnits from 'api/graph_data_units/useGraphDataUnits';
 
 // i18n
 import { useIntl } from 'utils/cl-intl';
@@ -21,7 +15,6 @@ import {
 
 // utils
 import { getFormattedNumbers } from 'components/admin/GraphCards/_utils/parse';
-import { isPage } from 'utils/helperUtils';
 
 // typings
 import {
@@ -38,30 +31,15 @@ export default function usePostsByTime({
   const { formatMessage } = useIntl();
   const [currentResolution] = useState(resolution);
 
-  const { pathname } = useLocation();
-  const { id: graphId } = useNode();
-  const isAdminPage = isPage('admin', pathname);
-  const { reportId } = useReportContext();
-
-  const { data: analyticsLive } = useLivePostsByTime(
-    {
+  const analytics = useGraphDataUnits<Response>({
+    resolvedName: 'PostsByTimeWidget',
+    queryParameters: {
       projectId,
       startAtMoment,
       endAtMoment,
       resolution,
     },
-    { enabled: isAdminPage }
-  );
-
-  const { data: analyticsPublished } = useGraphDataUnitsPublished<Response>(
-    {
-      reportId,
-      graphId,
-    },
-    { enabled: !isAdminPage }
-  );
-
-  const analytics = analyticsLive ?? analyticsPublished;
+  });
 
   const timeSeries = useMemo(
     () =>
